+++ src/game/enemy.ts
@@ -1,6 +1,7 @@
 /* Converted from TimePilot.Enemy.js (AMD) to ESM TypeScript. */
 import { levels, scoring } from "./constants";
 import userOptions from "./user-options";
+import { drawDebugVectors } from "./debug-vectors";
 import helpers from "./engine/helpers";
 import palette from "./palette";
 import { getDespawnRadius } from "./viewport";
@@ -23,6 +24,7 @@ class Enemy implements EnemyInstance {
   private _gameArena: GameArenaInstance;
   private _gameTicker: TickerInstance;
   private _player: PlayerInstance;
+  private _steeringHeading: Heading;
 
   isAlive = true;
   removeMe = false;
@@ -51,6 +53,7 @@ class Enemy implements EnemyInstance {
       ...options,
     };
     this._data.hitPoints = this.getLevelData().hitPoints;
+    this._steeringHeading = this._data.heading;
 
     this._enemySprite = new Image();
     this._enemySprite.src = this.getLevelData().sprite.src;
@@ -166,15 +169,32 @@ class Enemy implements EnemyInstance {
 
     this._checkInArena();
 
-    if (canTurn) {
-      let turnTo = helpers.findHeading(this._data, {
-        posX: player.posX + levelData.width / 2,
-        posY: player.posY + levelData.height / 2,
-      });
-      turnTo = Math.floor(turnTo / 22.5) * 22.5;
+    this._steeringHeading = this._getSteeringHeading(levelData, formationActive);
 
-      enemy.heading = helpers.rotateTo(turnTo, enemy.heading, 22.5);
+    if (canTurn) {
+      enemy.heading = helpers.rotateTo(
+        this._steeringHeading,
+        enemy.heading,
+        22.5
+      );
     }
+  };
+
+  private _getSteeringHeading = (
+    levelData: EnemyConfig,
+    formationActive: boolean
+  ): Heading => {
+    if (!levelData.tracksPlayer || formationActive || this.removeMe) {
+      return this._data.heading;
+    }
+
+    const player = this._player.getData();
+    const turnTo = helpers.findHeading(this._data, {
+      posX: player.posX + levelData.width / 2,
+      posY: player.posY + levelData.height / 2,
+    });
+
+    return Math.floor(turnTo / 22.5) * 22.5;
   };
 
   private _applyFormationWave = (tick: number): void => {
@@ -334,6 +354,25 @@ class Enemy implements EnemyInstance {
         levelData.hitRadius,
         {
           borderColor: palette.debug.enemyHitbox,
+        }
+      );
+    }
+
+    if (
+      userOptions.enableDebug &&
+      userOptions.debug.showHeadingVectors &&
+      !this._data.deathTick
+    ) {
+      const context = this._gameArena.getContext() as CanvasRenderingContext2D;
+
+      drawDebugVectors(
+        context,
+        this._data.posX - this._player.getData().posX,
+        this._data.posY - this._player.getData().posY,
+        this._data.heading,
+        this._steeringHeading,
+        {
+          fillTurnArc: userOptions.debug.showSteeringArc,
         }
       );
     }